+++ .github/workflows/test.yaml
@@ -85,8 +85,8 @@ jobs:
         with:
           name: ${{ matrix.os }}-${{ matrix.arch }}-libraries
           path: |
-            llm/build/**/bin/*
-            llm/build/**/*.a
+            payloads/build/**/*
+            payloads/build/**/*.a
   generate-cuda:
     needs: [changes]
     if: ${{ needs.changes.outputs.GENERATE_CUDA == 'True' }}
@@ -118,7 +118,7 @@ jobs:
         with:
           name: cuda-${{ matrix.cuda-version }}-libraries
           path: |
-            llm/build/**/bin/*
+            payloads/build/**/*
             dist/windows-amd64/**
   generate-rocm:
     needs: [changes]
@@ -151,7 +151,7 @@ jobs:
         with:
           name: rocm-${{ matrix.rocm-version }}-libraries
           path: |
-            llm/build/**/bin/*
+            payloads/build/**/*
             dist/windows-amd64/**
 
   # ROCm generation step
@@ -264,12 +264,12 @@ jobs:
           esac >>$GITHUB_ENV
         shell: bash
       - run: |
-          mkdir -p llm/build/linux/$ARCH/stub/bin
-          touch llm/build/linux/$ARCH/stub/bin/ollama_llama_server
+          mkdir -p payloads/build/linux/$ARCH/stub
+          touch payloads/build/linux/$ARCH/stub/ollama_llama_server
         if: ${{ startsWith(matrix.os, 'ubuntu-') }}
       - run: |
-          mkdir -p llm/build/darwin/$ARCH/stub/bin
-          touch llm/build/darwin/$ARCH/stub/bin/ollama_llama_server
+          mkdir -p payloads/build/darwin/$ARCH/stub
+          touch payloads/build/darwin/$ARCH/stub/ollama_llama_server
         if: ${{ startsWith(matrix.os, 'macos-') }}
       - uses: golangci/golangci-lint-action@v6
         with:
@@ -301,17 +301,17 @@ jobs:
           cache: true
       - run: |
           case ${{ matrix.arch }} in
-            amd64) echo ARCH=x86_64 ;;
+            amd64) echo ARCH=amd64 ;;
             arm64) echo ARCH=arm64 ;;
           esac >>$GITHUB_ENV
         shell: bash
       - run: |
-          mkdir -p llm/build/linux/$ARCH/stub/bin
-          touch llm/build/linux/$ARCH/stub/bin/ollama_llama_server
+          mkdir -p payloads/build/linux/$ARCH/stub
+          touch payloads/build/linux/$ARCH/stub/ollama_llama_server
         if: ${{ startsWith(matrix.os, 'ubuntu-') }}
       - run: |
-          mkdir -p llm/build/darwin/$ARCH/stub/bin
-          touch llm/build/darwin/$ARCH/stub/bin/ollama_llama_server
+          mkdir -p payloads/build/darwin/$ARCH/stub
+          touch payloads/build/darwin/$ARCH/stub/ollama_llama_server
         if: ${{ startsWith(matrix.os, 'macos-') }}
         shell: bash
       - run: go generate ./...
