--- llm/payloads_linux.go
@@ -1,6 +1,0 @@
-package llm
-
-import "embed"
-
-//go:embed build/linux/*/*/bin/*
-var libEmbed embed.FS
