+++ payloads/payloads_darwin_arm64.go
@@ -1,8 +1,8 @@
-package llm
+package payloads
 
 import "embed"
 
 // Darwin payloads separated by architecture to avoid duplicate payloads when cross compiling
 
-//go:embed build/darwin/arm64/*/bin/*
+//go:embed build/darwin/arm64/*/*
 var libEmbed embed.FS
