+++ payloads/payloads_linux.go
@@ -1,0 +1,6 @@
+package payloads
+
+import "embed"
+
+//go:embed build/linux/*/*/*
+var libEmbed embed.FS
